SOL-80176: switched IllegalStateException -> RuntimeException
Spring for some reason does NOT schedule late binding retries for Producer bindings if the exception during binding is an IllegalStateException or IllegalArgumentException (BindingService#assertNotIllegalException is where the check is done). This apparently does not apply to Consumer bindings. Since we want to schedule late bindings also for Producers in case of an inability to resolve Solace broker hostnames, the exception thrown from the proxy session must not be one of these filtered exceptions.